Using precreated native window
Tizen can provide a precreated window handle
to an application running as a candidate process.
Before creating the application's window, check whether
it is a precreated window handle and pass it to the embedder.
However, the candidate process on TV only applies to in-house applications
and is not supported by the FlutterApplication class,
which is currently based on CoreUIApplication.
Therefore, policy support should be given priority.

(This feature has been tested on RPI-based tizen profile (tizen9.0))